Добавляет base_url, разбивает константы на отдельные файлы
