+++ project/src/components/city-places-list/city-places-list.tsx
@@ -1,9 +1,10 @@
-import {getStringHashCode} from '../../utils/get-string-hash-code';
-import {PlaceCard} from '../place-card/place-card';
-import {PlacesSorting} from '../places-sorting/places-sorting';
+import {City} from 'settings/city';
+import {getStringHashCode} from 'utils/get-string-hash-code';
+import {PlaceCard} from 'components/place-card/place-card';
+import {PlacesSorting} from 'components/places-sorting/places-sorting';
 
 type CityPlacesListProps = {
-  cityName: string;
+  cityName: City;
   cityPlacesCount: number;
 };
 
@@ -15,7 +16,7 @@ function CityPlacesList({cityName, cityPlacesCount}: CityPlacesListProps): JSX.E
       <div className="cities__places-container container">
         <section className="cities__places places">
           <h2 className="visually-hidden">Places</h2>
-          <b className="places__found">{cityPlacesCount} places to stay in {cityName}</b>
+          <b className="places__found">{cityPlacesCount} places to stay in <span style={{textTransform: 'capitalize'}}>{cityName}</span></b>
           <PlacesSorting />
           <div className="cities__places-list places__list tabs__content">
             {Array.from(Array(MOCK_PLACE_CARD_COUNT)).map((_, index) => <PlaceCard id={index + 1} key={getStringHashCode(`${index}${cityName}`)} />)}
