+++ project/src/components/favorites-empty-screen/favorites-empty-screen.tsx
@@ -1,5 +1,5 @@
-import {Footer} from '../footer/footer';
-import {Header} from '../header/header';
+import {Footer} from 'components/footer/footer';
+import {Header} from 'components/header/header';
 
 function FavoritesEmptyScreen(): JSX.Element {
   return (
