--- project/src/const.ts
@@ -1,18 +1,0 @@
-enum AppRoute {
-  Favorites = '/favorites',
-  Login = '/login',
-  NotFound = '/404',
-  Property = '/offer',
-  Root = '/',
-}
-
-enum AuthorizationStatus {
-  Auth = 'AUTH',
-  NotAuth = 'NOT_AUTH',
-  Unknown = 'UNKNOWN',
-}
-
-export {
-  AppRoute,
-  AuthorizationStatus
-};
