+++ project/src/settings/app-route.ts
@@ -1,0 +1,9 @@
+enum AppRoute {
+  Favorites = '/favorites',
+  Login = '/login',
+  NotFound = '/404',
+  Property = '/offer',
+  Root = '/',
+}
+
+export {AppRoute};
