+++ project/src/settings/authorization-status.ts
@@ -1,0 +1,7 @@
+enum AuthorizationStatus {
+  Auth = 'AUTH',
+  NotAuth = 'NOT_AUTH',
+  Unknown = 'UNKNOWN',
+}
+
+export {AuthorizationStatus};
