+++ project/src/settings/city.ts
@@ -1,0 +1,10 @@
+enum City {
+  Paris = 'paris',
+  Cologne = 'cologne',
+  Brussels = 'brussels',
+  Amsterdam = 'amsterdam',
+  Hamburg = 'hamburg',
+  Dusseldorf = 'dusseldorf',
+}
+
+export {City};
